Fetch configuration from server

+++ android/app/src/main/java/org/pathcheck/covidsafepaths/exposurenotifications/dto/RNExposureInformation.java
@@ -6,9 +6,9 @@ import java.util.UUID;
 public class RNExposureInformation {
   private String id;
   private long date;
-  private long duration; // Minutes
+  private double duration; // Minutes
 
-  public RNExposureInformation(long date, long duration) {
+  public RNExposureInformation(long date, double duration) {
     this.id = UUID.randomUUID().toString();
     this.date = date;
     this.duration = duration;
